+++ system/src/utils/FalloutUtils.mjs
@@ -72,6 +72,11 @@ export default class FalloutUtils {
 		return Math.abs(game.time.worldTime - lastChange) > maxTimeSkip;
 	}
 
+	static foundryMinVersion(version) {
+		const majorVersion = parseInt(game.version.split(".")[0]);
+		return majorVersion >= version;
+	}
+
 	static getLocalizedSkillAttribute(skill) {
 		return game.i18n.localize(
 			`FALLOUT.AbilityAbbr.${skill.system.defaultAttribute}`
@@ -89,6 +94,14 @@ export default class FalloutUtils {
 		if (localizedName === nameKey) localizedName = skill.name;
 
 		return localizedName;
+	}
+
+	static getMessageStyles() {
+		const messageStyles = this.foundryMinVersion(12)
+			? CONST.CHAT_MESSAGE_STYLES
+			: CONST.CHAT_MESSAGE_TYPES;
+
+		return messageStyles;
 	}
 
 	static lbsToKgs(value) {